Add the Part system.
The Part system is an annotation based RegistryWrapper. It will find and register all classes annotated with @Part. Anything that is annotated with @Part should also extend BasePart, as this will provide all the default regsitry wrapping goodness. Furthermore, the Part system allows for easier data generation integration, whereby you gen data in the genData method provided by BasePart. The API is open, and as such extensions can be made to it if such is required by a child mod. An example of how to use the Part system can be seen the in the test package.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>